refactor(org2): split the crate root and agent-session modules
